odoo_mcp/odoo-mcp-agent/pyproject.toml
[project]
name = "odoo-mcp-agent"
version = "0.1.0"
description = "OpenAI Agent with Odoo MCP Server Integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents",
    "openai-agents-mcp",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
